refactor: update release name in flutter-release.yml
This commit updates the release_name field in the .github/workflows/flutter-release.yml file. The previous value was "WeConnect-IUI-mobile" followed by the branch name, but it has been changed to "flutter-calculator" followed by the branch name. This change ensures that the release name accurately reflects the purpose of the Flutter application.